refactor: move command out of actor
