+++ src/services/cluster/actors/mod.rs
@@ -1,6 +1,5 @@
 pub(super) mod actor;
 pub(crate) use replication::PeerState;
-pub mod command;
 mod listening_actor;
 
 pub(crate) mod peer;
