+++ src/services/cluster/mod.rs
@@ -2,3 +2,4 @@ pub(crate) mod actors;
 pub mod inbound;
 pub mod manager;
 pub mod outbound;
+pub mod command;
